requirements.txt
streamlit>=1.24.0
openai>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
reportlab